EnFlo virtual reality is here! Collision detection between any shapes with any orientations, with full 3D interactive viewing. Straight from CAD model (.stl binary format) into LabVIEW. No external libraries (no DLL-hell Tom :-D), all algorithms and derivations done from scratch and documented. Please do not modify any of the vis.
An Octree of user-specified resolution is built around the triangulated geometry and this is used to speed up the collision test from O(N) to O(log(N)). Every level down the tree eliminates 87.5% of the search subspace. A list of potential collisions is returned and then only the few necessary detailed triangle-triangle intersection tests are carried out. The point(s) of collision(s) are returned, and if there are no collisions the minimum separation vector and distance between the objects is returned if desired. The "Visualise STL file and Octree.vi" allows you to create and inspect the Octree and fly around it. Have a look at the example files.

Yet to be done: Housekeeping and integration with the rest of Enflo software. This is the purpose of the currently empty "EnFlo specific" folder. The eventual aim is to have a render window showing the tunnel, model(s), and traverse(s) all together. Collision points can be displayed if desired.

A more detailed description of what the vis do will follow once the integration with Enflo software is complete. The user should only ever need to use the vis in the root directory of this folder.