FINALLY cracked the convex hull algorithm! Ended up using the GiftWrap method instead of QuickHull. Although not O(nLog(n)) but O(nh), where n is the number of vertices and h is the number of hull points, it is a MUCH simpler algorithm that does not require face merging and horizon boundary evaluations(!). In brief - this algorithm wraps the surface of the Convex hull with triangles. These triangles are chosen such that all points are in the negative space of their hyperlanes. Every new triangle adds two new edges to the processing stack. Care must be taken to handle duplicated edges and degenerate triangles. Have play with the 3D example vi....
Anyway, with this working, we are now able to "extrude" our traverse! This means we can check a direct move for collisions in one go rather than over many steps, possibly missing small features. I have implemented this feature in the top level behemoth and tested the code with some made up moves and it works exactly as expected :-) Now we just need to make use of it in the route checking vi's. I have not modified any of those, just the low level re-link sub vi so it all continues to work as usual.