TCP/IP video streaming is here! New set of VIs to allow a client to initiate a USB camera to stream video over the network from the camera server and be viewed in real time. The stream can also be recorded locally to an avi.
Also in this commit is a VI that allows the addition of multi-line title or caption text to an existing image, with the options of changing font, size, style, colour and background colour.

To operate the TCP/IP VIs, the camera server needs to export the VI named "Video livestream (TCP SERVER).vi", and it needs to be loaded automatically at start-up. As for the client, the only VI to use is the one called "Start livestream video from CamServer (TCP CLIENT).vi".