Bug fixes to previous revision:
Added a new Extra command called !Trav_Profile_Plug_In which operates inside Trav_single_profile.vi so that the number of points visited in a profile can be controlled by the measured points so far. For example the vi Correlation_Profile.vi expects the profile to travel from high to low correlation coefficient and tries to give about 8 points over the correlation coefficient space from 1 to 0.1. The plug in essentially controls how many points are jumped in the profile and if the profile should be terminated.