New fun and hopefully useful tool! A 3D potential flow solver written entirely in LabVIEW for estimating and correcting traverse / probe blockage effects. Simply draw your setup in any 3D CAD program, export it in STL format and load into this tool. You can vary U_ref, and choose any point in space at which to compute the velocity vector. You can also perform virtual traverse profiles as well as create particle traces from any starting point. Surface Cp distribution can also be computed and shown visually, where you are free to fly around the model and enjoy the pretty colours. Once the geometry has been loaded and aolved for, you can turn off the action selector switches on the left and just do, for example, point velocity measurements or traverse profiles (which are very quick, rather than re-solving the system each time). If you change U_ref or the geometry, then of course you must re-compute the matrix and solve it. The default settings should work well for most problems. Pop-ups will appear to show the end-user what is going on, along with the option to stop the particular step, rather than just leaving them wondering whether the program has hung.
What this code currently cannot do: Lifting bodies (i.e. those which require a Kutta condition, or circulation). This requires a wake model in the form of shedding panels. For typical probe-like objects (cylinders and hemispheres etc.), this is not really necessary when estimating the upstream influence.
Also, not applicable in flows where viscous effects are significant such as thick or separating boundary layers, shear flows, wake flows and reversing flows. Again, not an issue when upstream of the objects of interest.
The freestream cannot be turbulent. If your probe is submerged in a boundary layer, then use the local mean velocity to get an estimate of the average effect.

Another thing to watch out for - CAD programs output rather nasty STLs for CFD, they are optimised for minimum number of triangles rather than the most equilateral ones. Sometimes the solution will fail if it is so bad. Some kind of mesh refinement tool would be helpful here, or a program that outputs nicer STLs (any ideas?). Sometimes going for the lower resolution STL file will produce much better results.

Final caveat - this being a vortex lattice method (which is the lowest order method), the flow-field very near the body will appear to have wiggles in between the panel edges. This is because the flow boundary condition is only satisfied exactly at one point on the panel. The effect can be mitigated by having smaller panels. However it has no effect on the accuracy of the Cp distribution or computation of the induced velocity in the far-field.

Reality checks: look for Cp_max being close to 1, and Cp_min not being much less than -6 or so in most cases. Code was validated using a cylinder with aspect ratio 1000 (effectively 2D) and gave Cp_max = 0.999 and Cp_min = -2.97 (theoretical values are 1.0 and -3.0 respectively). Next with a sphere which gave Cp_max = 0.998 and Cp_min = -1.279 (theoretical values are 1.0 and -1.25 respectively). So errors in Cp_min are 1 - 4%. Note that Cp_min on the sphere was taken at a location where the triangles were not so skewed. Results can be exported from the graphs as desired, or the VI can be called from a higher level and used to generate even more data!

Enjoy :-)